requirements.txt
cmudict
cn2an
faster-whisper==0.10.1
g2p_en
GPUtil
gradio
jieba
langid
librosa==0.9.2
loguru
matplotlib
num2words
numba
numpy
psutil
pyannote.audio>=3.1.0
pyloudnorm
# pyopenjtalk-prebuilt  # Should be manually uninstalled
pyopenjtalk-dict
pypinyin
# pyworld  # Not supported on Windows without Cython...
PyYAML
requests
safetensors
scipy
tensorboard
torch>=2.1,<2.2
transformers
umap-learn
runpod
